data-preview/requirements.txt
Flask==2.3.1
Flask-MySQL==1.5.2
flask-mysql-connector==1.1.0
Flask-SQLAlchemy==3.0.3
mysql-connector-python==8.0.33
pandas==2.0.1
plotly==5.14.1
PyMySQL==1.0.3
python-dotenv==1.0.0
SQLAlchemy==2.0.10
sqlparse==0.4.4